moved tuple prior to dedicated file
