+++ autofit/mapper/prior/prior.py
@@ -11,14 +11,8 @@ from scipy.special import erfcinv
 
 from autoconf import conf
 from autofit import exc
-from autofit.mapper.model_object import ModelObject
 from autofit.mapper.prior.arithmetic import ArithmeticMixin
 from autofit.mapper.prior.deferred import DeferredArgument
-from autofit.mapper.prior_model.attribute_pair import (
-    cast_collection,
-    PriorNameValue,
-    InstanceNameValue,
-)
 from autofit.mapper.variable import Variable
 
 
@@ -29,90 +23,6 @@ class Limits:
             cls, [attribute_name, "gaussian_limits"]
         )
         return limit_dict["lower"], limit_dict["upper"]
-
-
-class TuplePrior(ModelObject):
-    """
-    A prior comprising one or more priors in a tuple
-    """
-
-    @property
-    @cast_collection(PriorNameValue)
-    def prior_tuples(self):
-        """
-        Returns
-        -------
-        priors: [(String, Prior)]
-            A list of priors contained in this tuple
-        """
-        return list(filter(lambda t: isinstance(t[1], Prior), self.__dict__.items()))
-
-    @property
-    def unique_prior_tuples(self):
-        return self.prior_tuples
-
-    @property
-    @cast_collection(InstanceNameValue)
-    def instance_tuples(self):
-        """
-        Returns
-        -------
-        instances: [(String, instance)]
-            A list of instances
-        """
-        return list(
-            sorted(
-                filter(lambda t: isinstance(t[1], float), self.__dict__.items()),
-                key=lambda tup: tup[0],
-            )
-        )
-
-    def value_for_arguments(self, arguments):
-        """
-        Parameters
-        ----------
-        arguments: {Prior: float}
-            A dictionary of arguments
-
-        Returns
-        -------
-        tuple: (float,...)
-            A tuple of float values
-        """
-
-        def convert(tup):
-            if hasattr(tup, "prior"):
-                return arguments[tup.prior]
-            return tup.instance
-
-        return tuple(
-            map(
-                convert,
-                sorted(
-                    self.prior_tuples + self.instance_tuples, key=lambda tup: tup.name
-                ),
-            )
-        )
-
-    def gaussian_tuple_prior_for_arguments(self, arguments):
-        """
-        Parameters
-        ----------
-        arguments: {Prior: float}
-            A dictionary of arguments
-
-        Returns
-        -------
-        tuple_prior: TuplePrior
-            A new tuple prior with gaussian priors
-        """
-        tuple_prior = TuplePrior()
-        for prior_tuple in self.prior_tuples:
-            setattr(tuple_prior, prior_tuple.name, arguments[prior_tuple.prior])
-        return tuple_prior
-
-    def __getitem__(self, item):
-        return self.prior_tuples[item][1]
 
 
 class Prior(Variable, ABC, ArithmeticMixin):
